+++ src/main/java/raven/datetime/TimePicker.java
@@ -1,7 +1,6 @@
 package raven.datetime;
 
 import com.formdev.flatlaf.FlatClientProperties;
-import com.formdev.flatlaf.extras.FlatSVGIcon;
 import net.miginfocom.swing.MigLayout;
 import raven.datetime.component.PanelPopupEditor;
 import raven.datetime.component.time.Header;
@@ -14,6 +13,7 @@ import raven.datetime.event.TimeSelectionEvent;
 import raven.datetime.event.TimeSelectionListener;
 import raven.datetime.util.InputUtils;
 import raven.datetime.util.InputValidationListener;
+import raven.datetime.util.Utils;
 
 import javax.swing.*;
 import java.awt.*;
@@ -230,7 +230,7 @@ public class TimePicker extends PanelPopupEditor implements TimeSelectionModelLi
 
     private void installEditor(JFormattedTextField editor) {
         JToolBar toolBar = new JToolBar();
-        editorButton = new JButton(editorIcon != null ? editorIcon : new FlatSVGIcon("raven/datetime/icon/clock.svg", 0.8f));
+        editorButton = new JButton(editorIcon != null ? editorIcon : Utils.createIcon("raven/datetime/icon/clock.svg", 0.38f));
         toolBar.add(editorButton);
         editorButton.addActionListener(e -> {
             if (editor.isEnabled()) {