+++ src/main/java/raven/datetime/component/date/Header.java
@@ -1,11 +1,11 @@
 package raven.datetime.component.date;
 
 import com.formdev.flatlaf.FlatClientProperties;
-import com.formdev.flatlaf.extras.FlatSVGIcon;
 import net.miginfocom.swing.MigLayout;
 import raven.datetime.DatePicker;
 import raven.datetime.component.date.event.DateControlEvent;
 import raven.datetime.component.date.event.DateControlListener;
+import raven.datetime.util.Utils;
 
 import javax.swing.*;
 
@@ -67,11 +67,11 @@ public class Header extends JPanel {
     }
 
     protected Icon createDefaultBackIcon() {
-        return new FlatSVGIcon("raven/datetime/icon/back.svg");
+        return Utils.createIcon("raven/datetime/icon/back.svg", 1f);
     }
 
     protected Icon createDefaultForwardIcon() {
-        return new FlatSVGIcon("raven/datetime/icon/forward.svg");
+        return Utils.createIcon("raven/datetime/icon/forward.svg", 1f);
     }
 
     public void addDateControlListener(DateControlListener listener) {